vit_service/requirements.txt
torch==1.12.0+cpu
torchvision==0.13.0+cpu
boto3
fastapi[all]==0.88.0
gunicorn==20.1.0
Pillow==9.5.0